// packages/frontend-common/formats/chart/hierarchical-tree/type.ts
export type Datum = {
    source: string;
    source_title?: string;
    source_description?: string;
    source_detail1?: string;
    source_detail2?: string;
    source_detail3?: string;

    target: string;
    target_title?: string;
    target_description?: string;
    target_detail1?: string;
    target_detail2?: string;
    target_detail3?: string;

    weight?: number;
};